Refactoring №2. Поправил ошибку отоброжением ЗП не указана
